Revert "Css fix julie"
