+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -14,6 +14,5 @@ class VerifyCsrfToken extends Middleware
     protected $except = [
         //
         'interets/update_Interets/*',
-        
     ];
 }
