+++ app/Models/Friendship_Request.php
@@ -7,8 +7,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Friendship_Request extends Model
 {
-    protected $table= 'friendships_requests';
-    protected $fillable = ["id_user_send", "id_user_receive", "status"];
+    protected $table= 'Friendships_Requests';
+    protected $fillable = ["id_user_send", "id_user_receive", "state"];
     public $timestamps = false;
 
     public static function AddFriendRequest($user1, $user2) {
@@ -21,7 +21,7 @@ class Friendship_Request extends Model
                 $request = [
                     "id_user_send" => $user1,
                     "id_user_receive" => $user2,
-                    "status" => 'pending'
+                    "state" => 'Sent'
                 ];
                 Friendship_Request::Create($request);
             }
@@ -31,7 +31,7 @@ class Friendship_Request extends Model
     public static function IsRefuse($user1, $user2) {
         $receive = Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2);
         if ($receive->count() > 0) {
-            if ($receive->get()->first()->status == "refused") {
+            if ($receive->get()->first()->state == "Refused") {
                 return true;
             }
         }
@@ -40,7 +40,7 @@ class Friendship_Request extends Model
 
     public static function GetFriendRequestReceive($userId) {
         $requests = [];
-        $requestsData = Friendship_Request::where("id_user_receive", $userId)->where("status", "pending")->get();
+        $requestsData = Friendship_Request::where("id_user_receive", $userId)->where("state", "Sent")->get();
         if ($requestsData->count() > 0) {
             foreach ($requestsData as $requestData) {
                 $user = User::where("id", $requestData->id_user_send)->get()->first();
@@ -49,9 +49,9 @@ class Friendship_Request extends Model
         }
         return $requests;
     }
-    public static function GetFriendRequestpending($userId) {
+    public static function GetFriendRequestSent($userId) {
         $requests = [];
-        $requestsData = Friendship_Request::where("id_user_send", $userId)->where("status", "pending")->get();
+        $requestsData = Friendship_Request::where("id_user_send", $userId)->where("state", "Sent")->get();
         if ($requestsData->count() > 0) {
             foreach ($requestsData as $requestData) {
                 $user = User::where("id", $requestData->id_user_receive)->get()->first();
@@ -61,19 +61,19 @@ class Friendship_Request extends Model
         return $requests;
     }
 
-    public static function GetUserRelationstatus($user1, $user2) {
-        $status = Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1)->where("status", "pending");
+    public static function GetUserRelationState($user1, $user2) {
+        $state = Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1)->where("state", "Sent");
         $message = null;
-        if ($status->count() > 0) {
+        if ($state->count() > 0) {
             $message = "receive";
             return $message;
         }
-        $status = Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2);
-        if ($status->count() > 0) {
-            if ($status->get()->first()->status == "refused") {
+        $state = Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2);
+        if ($state->count() > 0) {
+            if ($state->get()->first()->state == "Refused") {
                 $message = "refuse";
-            } else if ($status->get()->first()->status == "pending") {
-                $message = "pending";
+            } else if ($state->get()->first()->state == "Sent") {
+                $message = "sent";
             }
         }
 
@@ -81,27 +81,31 @@ class Friendship_Request extends Model
     }
 
     public static function AcceptFriendRequest($user1, $user2) {
-        $status = ["status" => "accepted"];
-        Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->update($status);
+        $state = ["state" => "Accepted"];
+        Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->update($state);
     }
     public static function RefuseFriendRequest($user1, $user2) {
         $request = Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1);
         if ($request->count() > 0) {
-            $status = ["status" => "refused"];
-            Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1)->update($status);
+            $state = ["state" => "Refused"];
+            Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1)->update($state);
         }
     }
 
     public static function CancelFriendRequest($user1, $user2) {
         $other_request = Friendship_Request::where("id_user_send", $user2)->where("id_user_receive", $user1)->get();
         if ($other_request->count() > 0) {
-            if (!$other_request->first()->status == "refused") {
+            if (!$other_request->first()->state == "Refused") {
                 Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->delete();
                 return;
             }
         } else {
             Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->delete();
         }
+        /*$other_request = Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->get();
+        if (!($other_request->count() > 0 && $other_request->first()->state == "Refused")) {
+            Friendship_Request::where("id_user_send", $user1)->where("id_user_receive", $user2)->delete();
+        }*/
     }
 
     public static function RemoveFriendRequest($user1, $user2) {
