+++ composer.json
@@ -25,10 +25,7 @@
             "App\\": "app/",
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
-        },
-        "files": [
-            "app/Helpers/MeetupHelper.php"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
