+++ public/mix-manifest.json
@@ -8,8 +8,6 @@
     "/js/home-Overlay-login.js": "/js/home-Overlay-login.js",
     "/js/flash.js": "/js/flash.js",
     "/js/messages.js": "/js/messages.js",
-    "/js/layout.js": "/js/layout.js",
-
     "/css/profile.css": "/css/profile.css",
     "/css/app.css": "/css/app.css",
     "/css/meetup.css": "/css/meetup.css",
