+++ resources/css/app.css
@@ -692,38 +692,6 @@ font-size: 0.3em !important;
     -moz-user-select: none; /* Old versions of Firefox and Internet Explorer */
     user-select: none; /* Non-prefixed version, currently supported by Chrome, Edge, Opera and Firefox */
 }
-.btn_back {
-    position: absolute;
-    top: 0.3rem;
-    left: 6rem;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    background-color: #6795c68d;
-    color: white;
-    border: none;
-    border-radius: 1.25rem;
-    padding: 0.5rem 1rem;
-    font-size: 1rem;
-    font-weight: bold;
-    cursor: pointer;
-    transition: background-color 0.3s, transform 0.2s;
-    box-shadow: 0 0.25rem 0.375rem rgba(0, 0, 0, 0.1);
-    z-index: 100;
-}
-
-.btn_back .material-symbols-outlined {
-    font-size: 1.25rem;
-}
-
-.btn_back:hover {
-    background-color: var(--primary-400);
-    transform: scale(1.05);
-}
-
-.btn_back:active {
-    transform: scale(0.95);
-}
 
 .no_wrap {
     white-space: nowrap !important;
